auth-service/requirements.txt
Flask==3.0.3
gunicorn==22.0.0
SQLAlchemy==2.0.34
psycopg2-binary==2.9.9
alembic==1.13.2
PyJWT==2.9.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.1

